refactor: domain package 구조 변경
- order package 생성
  - Order, OrderResult 패키지화
- Promotion, PromotionProduct를 product 패키지에 포함
