+++ src/main/java/store/domain/product/Promotion.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.domain.product;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
