+++ src/test/java/store/view/parser/OrderInputParserTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.domain.Order;
+import store.domain.order.Order;
 import store.exception.InvalidInputException;
 
 class OrderInputParserTest {
